requirements.txt
duckduckgo-search
pdfplumber
reportlab
scipy
statsmodels
seaborn
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
pandas
numpy
matplotlib
plotly
scikit-learn
sqlalchemy
python-dotenv
openpyxl
fastapi
uvicorn
faiss-cpu
tabulate
python-pptx
